Cleaned up code logic in traversal functions
In depthFirst and breadthFirst search functions for both dGraph
and uGraph, I cleaned up the structure of some of the statements,
mostly having to do with the graph traveler. The overall logic is
still the same, but there were some unneccessary if-statements
and unused variables that got cleaned up. All changes were tested
and confirmed correct with the respective testing suites.