Setting up graph generation code and testing changes.
Set up the GraphGen directory to contain the code for generating
large numbers of arbitrarily sized strongly-connected graphs. I
will use this to generate a large database of graphs to perform
testing upon. Also made some changes to the randomization functions
for the edge-weighting in the unit tests, this more evenly distributes
the weights, which makes our tests run slower but represents a more
accurate and complex scenario for testing.